add timecalc
